UZEM: Merged in the rearranged CPU instruction decoder switch tree.
Important:

It seems like there are bugs in the instruction timing calculations, see
the TODO comments. There may be more locations of interest! For now I only
attempted to merge as-is, not attempting to fix these, just adding a few
comments here and there to note where the location of the cycles variable
is important.