UZEM: Timer1's code was modified for performance.
What was introduced is the calculation of time between events of Timer1, so
update_hardware can get Timer1 done with a single comparison if it is still
far from reaching the next compare or overflow event.

The logic of Timer1 wasn't changed (except for a removal of a TCNT1 assignment
from write_io, which is unnecessary since these 16 bit variables are only set
up and used within the Timer1 logic). No bugs except from imperfections in the
time between events calculation should appear.

Note that with this, write_io necessarily has to clear the new timer1_next
variable whenever it writes to a timer register under the code avoided by the
new logic. For this, new cases had to be added for the OCR1A and OCR1B
registers.