UZEM: Small tweaks ending up in a massive performance bump!
There are four distinct changes in this package, which performed individually
for some reason produce dubious results in performance, but combined, give a
massive near 30 percents increase!

Change 0 relates to Timer 1 again. The TCNT1 counter is no longer incremented
on the fast path within update_hardware, but rather calculated on demand. The
rationale is that it saves a load, an increment, and a store on the critical
path.

Change 1 concerns how the line buffer is filled. Now, taking advantage of that
it is possible to use the line buffer as a circular buffer, dropping the
current_cycles variable, the cycleCounter is used directly. This, like the
change for the timer, removes a load, an increment, and a store from the
critical path.

Change 2 moves the delayed outputs off of update_hardware, to
update_hardware_ins, along with placing a note there on why it can be done. It
can be done since as of now, all AVR instructions which write to I/O ports do
this in their last cycle, so a delayed output can only appear then, which is
so handled correctly within update_hardware_ins. By alone this has a meager
benefit (eliminates a load, and a single hardly ever taken branch), but rather
prepares for Change 3.

Change 3 creates an inline stub for update_hardware named
update_hardware_fast, which can perform the fast path inlined. This is then
used with the most common instructions (like ld, st, lpm, mul, rjmp, ijmp, the
latter heavily used in code block based graphics modes), and also on the end
of the instruction decoder. This saves function call overheads, one for all
instructions, and one extra for those frequently used 2 cycle instructions.
This is what actually gives the performance bump, but it doesn't work without
the previous changes cleaning up the fast path.