MODE74: Uzebox logo added.
From now it is possible to use an Uzebox intro logo like in other video modes
using the INTRO_LOGO define. This is designed so it interferes the least with
the flexibility of the mode (that is, the logo should simply work fine with
most sane use-cases).

The two examples are also updated to show the logo.