MODE74: Reset on frame return capability added.
The reset on frame end capability allows for creating a main program which is
only allowed to run until the start of the next video frame. Then it is
terminated. Upon the return from the video frame, it is restarted with an
empty stack.

The sprite engine received a few necessary fixes to work properly with this
feature, that is, it can be terminated within render.

A new example is added, the same sprite demo, but modified to this reset
enabled concept.

The reset function was also tested succesfully with the SD loading feature in
the Multicolor mode.

A major benefit of this feature, and the default configuration is set up
accordingly is that almost the complete Stack can be eliminated: the main
program can use the Palette buffer as Stack. This saves 240 bytes in the
0x1010 - 0x10FF range (the 0x1000 - 0x100F range is used as stack for the
video mode, which is sufficient for this purpose).

Later the video mode stack might be moved off to an arbitrary location so the
Palette buffer and Main program stack may be moved up into the 0x1000 - 0x10FF
region.