MODE74: Color index 0 disable feature added.
Color index 0 disable if enabled causes omitting the fill of color 0 in the
palette buffer, thus saving 16 bytes of RAM. Color 0 is the transparent color
for sprites, so it might be likely to be left unused.

This disables the "normal" 2bpp mode since it uses color indices 0 - 3. The
Multicolor 2bpp mode is still available if enabled (it may be used with
repeating a single VRAM row to conserve RAM memory if necessary for a simple
2bpp screen).