MODE74: Minor optimizations.
The pixel output got slightly faster by a rearrangement of the RAM tile
allocator allowing for using less registers.

The reset feature got two small functions to add a halt and a sequence point
to routines meant to be terminated by a new frame. This gives cleaner code for
those while conforming the C standard.