KERNELBOOST: 57 cycles gained on HSync entry, giving ~2000 extra VBlank cycles
The modification replaces the jitter elimination part in the HSync and
VSync entry codes to a more efficient solution.

It is not fully compatible with the original kernel, but mostly works.
Notably all timing is shifted 57 cycles "down" relative to the timer,
which normally isn't a problem (expect for timer value related comments).
However video modes using the timer for terminating their scanlines (such
as Mode 13) doesn't work with this modification! (their timer related
logic has to be modified to follow the 57 cycles shift)