Added p2m (polyphonic to monophonic) MIDI tool
This is an open-source python tool (not written by me) that converts a
midi file containing a polyphonic track into a multitrack one (one voice
per track).

This is very useful if you've recorded a Uzebox track using a MIDI
keyboard and you pressed multiple keys at once. Just use the same
instrument for the multiple channels, and now you can fake a bit of
polyphony support on the Uzebox.