MODE74: Polished example programs.
Configurations are now in a kconf.h file (instead of m74cfg.h suggesting
only Mode 74 specifics), this is more appropriate for this video mode.

Makefiles also got a little polish to get them slightly more readable.
EEPROM generation was removed (it is not useful since the EEPROM is
normally shared on an Uzebox between games).