Simple kernel boost (#78)
* KERNELBOOST: 57 cycles gained on HSync entry, giving ~2000 extra VBlank cycles

The modification replaces the jitter elimination part in the HSync and
VSync entry codes to a more efficient solution.

It is not fully compatible with the original kernel, but mostly works.
Notably all timing is shifted 57 cycles "down" relative to the timer,
which normally isn't a problem (expect for timer value related comments).
However video modes using the timer for terminating their scanlines (such
as Mode 13) doesn't work with this modification! (their timer related
logic has to be modified to follow the 57 cycles shift)

* Kernel boost: Added Timer1 terminated video mode compatibility.

A new TIMER1_DISPLACE constant is introduced which can be used in timer
terminated video modes to make them compatible with different kernels. Its
use is added to Mode 13, so now it compiles correctly with the kernel
boost enabled.

* Kernel boost: Cleaning up.

The Vsync generation also became smaller and a little faster.

* Kernel boost: Cleaned up Mode 13's modification.

It now can be used as a general working example for setting up other
Timer1 terminated video modes.