Mode3 new frame driver (#79)
* MODE3: Converted assembly file to Unix new lines.

Preparation for a larger modification.

* MODE3: Blitter optimizations.

The blitter main loop became about 25% faster.

* MODE3: Removed old commented-out blitter code.

* MODE3: Minor overhead optimizations in the blitter

Made the blitter slightly smaller and faster, using one byte less stack.

* MODE3: Minor overhead and loop optimizations in the blitter.

Made it sligthly faster and using a byte less stack (no stack usage now).

* MODE3: Further blitter loop optimization.

An earlier quirk of the loop's algorithm was removed, making it both
smaller and faster.

* MODE3: Further blitter loop optimization.

An earlier quirk of the loop's algorithm was removed, making it both
smaller and faster.

* MODE3: Converted videoMode3.c to Unix format.

Preparing for a larger change.

* MODE3: Rewrote ProcessSprites().

The new function body should be sligthly faster and should contain less
bugs while providing better support for irregular tile sizes (not 8x8).

Also added copyright notices where appropriate.

* MODE3: RestoreBackground reimplemented in assembly.

This saves a few hundred clocks for a frame as the assembly generated from
the C code was poor.

* MODE3: Made RestoreBackground a little smaller and faster.

* MODE3: Converted videoMode3.h and videoMode3.def.h to Unix newlines.

In preparation for larger modifications.

* MODE3: Added SPRITES_VSYNC_PROCESS compile time flag.

By default the flag's value is 1, which is the behavior of the kernel
before the introduction of this flag. By setting it to 0, it is possible
to turn off automatic sprite management.

* MODE3: Added SPRITES_AUTO_PROCESS compile time flag.

Setting this zero allows for using a simple blitter instead of Mode3's
default sprite engine.

* MODE3: Added RAM Tile restore list disable option.

This compile time option can be used to remove the RAM Tile restore list,
useful for games not needing it (which only use User RAM tiles or always
rebuild the full VRAM).

* MODE3: Added RT_ALIGNED compile time option.

With this, you can get RAM tiles aligned to 64 byte boundary, slightly
improving blitting performance. The RAM tiles this case are located after
the VRAM which is aligned already to avoid wasting bytes.

* MODE3: Documenting critical parts.

Preparations for understanding and rewriting the frame code of the mode
(for the later 5.5 cycles / pixel option, and also cleaning it up as much
as possible).

* MODE3: New scanline loop and higher resolution feature.

The normal scrolling Mode 3 can now have 30 tiles width. The
RESOLUTION_EXT compile time flag is added to enable compiling the 5.5
cycles / pixel scanline loop instead of the 6 cycles / pixel one, if
that is used, scrolling Mode 3 can have up to 31 tiles width (not more
only due to the VRAM layout).

In the next phase, the non-scrolling Mode 3 will also be transformed to
have these capabilities (the 5.5 cycles / pixel scanline loop supporting
up to 32 tiles width).

* MODE3: Added non-scrolling support.

Now it works both with scrolling and non-scrolling games, with the default
settings everything should compile and run fine as before. RESOLUTION_EXT
is completely transparent.

* MODE3: Fixing accidental commit with RESOLUTION_EXT enabled by default.

It should be disabled to produce compatible visual output with the
original Mode 3, I had it set for testing.

* MODE3: Miscalculated wait corrected.

The modified wait wasn't updated to the modified count of instructions on
the oter branch.

* MODE3: Added kernel timing compatibility constant.

The TIMER1_DISPLACE constant introduced with the kernel boost is applied
on the new Mode3 frame driver.