MODE3: Fixing sprite Y scroll. SPRITE_OFF flag added.
The sprite Y scroll of the scrolling Mode 3 should now work in a
reasonable manner (sprite displays when vertically it is in the visible
region, no artifacts present when it is outside).

A SPRITE_OFF flag is added. This flag can be used on sprites to disable
them (so you don't need to destroy a sprite coordinate to move it off
screen).