Bootloader API library added with a demo.
This API library enables using SDHC cards and FAT32 filesystems with
fragmentation through the V.5.x.x bootloader if it is available (otherwise
reverting to an MMC and FAT16 implementation).

The demo shows off the basic features of this library, also doubling as an
useful program for querying SD cards.