Mixer boost (#88)
A significantly faster inline mixer based on Flight of a Dragon and CunningFellow's mixer optimizations. It doesn't change anything in the behaviour, the most notable benefit is more CPU time available for user code in VBlank, also in some video modes, it may enable using a more complex inline mixer.