MODE74: Added noreturn attribute to M74_Halt() improving generated code.
The M74_Halt() function is used when Mode 74 is used with frame reset, to
wait for the next frame in an infinite loop. The next frame would then
jump to a reset address.

This function never returns. Adding __attribute__((noreturn)) indicates
this for GCC, allowing generating smaller code (for example a no-return
path doesn't need to restore registers, saving lots of pushes and pops).