MODE748: Display Disabled color added.
This allows for setting display color when it is disabled (which only
disables the mode's output and SPI RAM access, not affecting the length
of the frame interrupt). It could be useful for non-black transitions
where more significant SPI accesses are needed.

Also removed the leftover Uzebox logo output which doesn't work. Later
when it is fixed, it may be readded.