MODE748: Documentation updates & removed intro logo leftovers.
There will be no intro logo display in this mode, since the mode's most
common use case is combining it with heavy SD access. An intro logo can be
loaded off from an external file to save precious ROM space for things
which have to be in ROM (such as background tilesets).