Bugfix: videoMode3core.s, SetFont
Originally mentioned in: http://uzebox.org/forums/viewtopic.php?f=3&t=9994&p=29736&hilit=SetFont#p29736

In the line with add r20,21 (1535) the "r" is missing from "21". It should be "add r20,r21". It appears that 21 and r21 act the same though. Still, this is technically a typo and should be updated.