Bootloader workaround: Mac OS X generated files no longer show up as junk entries in the Game Selector.
This also means that neither valid UZE files beginning with an underscore
character would show up as by FAT16 filename only this can be used to
isolate those OS X files. This shouldn't be a problem as there is no known
game in existence with such a name (it is an extremely unlikely use case).