update peripheral support to work
Modified "mouse" byte to be "psupport" which is bit-masked with all peripherals the game supports. This field was previously unused by packrom or any emulator, though there were placeholders

Also used 1 reserved byte to represent the default/ideal starting "plugged in" state for the emulator, ie:
mouse=support
keyboard=default
multitap=support
esp8266=default

The above would cause emulation to start with the Keyboard and ESP8266 enabled by default. The emulator can display that Mouse and Multitap are supported, and they may optionally turn on those features with appropriate hotkeys. Some options that are nice to support, may be non-intuitive to casual users otherwise, so this simplifies things, optionally.

This modification matches the CUzeBox modifications which will start certain ROMs with certain "ideal" peripherals already turned on. Basically this gives the "they plugged it in" assumption of user intent, otherwise missing in emulation without reading up specific hotkeys(or on platforms where such hotkeys may not be practical).