Added UzeROM peripheral support+emulator default
The previous "mouse" option was not utilized. Converted this to "psupported" to represent all the peripherals the game supports(as info to the user).

Used 1 reserved byte to represent "pdefault", which is default peripherals to enable in the emulator(saving a hotkey from the user). This will simplify things for the user for the most common setups. They can still enable supported(but not default) peripherals using the standard hotkeys. This bridges the gap between hardware "they plugged it in" user intention, and the emulator which will generally start with a keyboard, mouse, internet, but not necessarily want to emulate them to the ROM by default.