Modified video core to support slow SNES controllers
I added a new game controller query routine. This solves problems with many of these cheap SNES controller core which have no shift registers but a Microcontroller onboard. These are often too slow to be queried by the stock readbuttons routine.

Use the define SLOW_CONTROLLERS=1 for enabling the new routine, which is default in this fork.

How it works:

The stock routine reads the controller during VSync with nearly no delays. This results in very short load and clock pulses, which are too fast for some controllers.
The new routine reads one bit per scanline, after the HSync pulse. It therefor needs 17 scanlines. These fit in the area below the VSYNC and the active video frame. Let's call it the upper border. This results in a much slower clock sequence for the controller without wasting time in delay loops.

Drawbacks:

At the moment there ist no mouse support because the upper border area has not eneough lines for that. Maybe I can move the routine to the active video area, but that might break video timing at some games.